Memory: semicolons as comments, signed integer operand

+++ Sources/QCPU/Shared/MemoryComponent/CompiledStatement.swift
@@ -33,6 +33,6 @@ extension MemoryComponent {
         }
 
         let instruction: Instruction
-        let operand: Int8?
+        let operand: Int?
     }
 }
